+++ test/GSqlQuery.SQLServer.Test/Data/Tables/Actor.cs
@@ -4,7 +4,7 @@ namespace GSqlQuery.SQLServer.Test.Data.Tables
     public class Actor : EntityExecute<Actor>
     {
         [Column("actor_id", Size = 5, IsAutoIncrementing = true, IsPrimaryKey = true)]
-        public long ActorId { get; set; }
+        public int ActorId { get; set; }
 
         [Column("first_name", Size = 45)]
         public string FirstName { get; set; }
@@ -18,7 +18,7 @@ namespace GSqlQuery.SQLServer.Test.Data.Tables
         public Actor()
         { }
 
-        public Actor(long actorId, string firstName, string lastName, DateTime lastUpdate)
+        public Actor(int actorId, string firstName, string lastName, DateTime lastUpdate)
         {
             ActorId = actorId;
             FirstName = firstName;
